DataType converted to long.

+++ Basics/Recursion/DSA.Basics.SumOfDigits/Program.cs
@@ -1,7 +1,7 @@
 ﻿using DSA.Basics.SumOfDigits;
 
 Console.Write("Enter a number greater than or equal to zero : ");
-int number = Convert.ToInt32(Console.ReadLine());
+long number = Convert.ToInt64(Console.ReadLine());
 
 Console.WriteLine(SumOfDigits.CalculateSumOfDigits(number));
